moe_utils cleanup after
